add upper bound to eval; eval stat collection
statistics on the results of node evaluation
are collected. these are broken down by the
evaluation type and the search depth

added a new upper bound to the evaluation
function. this rarely seems to be helpful.
it might be good to either tighten the
bound substantially or rip it out.

had to rearrange some stuff wrt constants
and types (this is a bit of a mess)